create module slide.js
