+++ src/animation.js
@@ -10,81 +10,6 @@ export function scrollIntoView(target){
     });
 }
 
-/**
- * Slide
- */
-
-export function slideUp(target, duration = 500, fn){
-    //this.log(`[animate] slide up`, arguments);
-
-    // before
-    setCSS(target, {
-        boxSizing: 'border-box',
-        height: target.scrollHeight + 'px'
-    });
-    setTransition(target, duration);
-
-    // animate
-    setCSS(target, {
-        height: '0px',
-        marginTop: '0px',
-        marginBottom: '0px',
-        paddingTop: '0px',
-        paddingBottom: '0px',
-    });
-
-    // end
-    setTimeout(() => {
-        setCSS(target, {
-            display: 'none',
-            height: ''
-        });
-        removeTransition(target);
-
-        // callback
-        if(typeof fn === 'function') fn();
-    }, duration);
-}
-
-export function slideDown(target, duration = 500, fn){
-    //this.log(`[animate] slide down`);
-
-    // before
-    setCSS(target, {
-        boxSizing: 'border-box',
-        height: '0px',
-        display: 'block'
-    });
-    setTransition(target, duration);
-
-    // animate
-    setCSS(target, {
-        height: `${target.scrollHeight + getElementMarginHeight(target)}px`,
-        marginTop: '',
-        marginBottom: '',
-        paddingTop: '',
-        paddingBottom: '',
-    });
-
-    // end
-    setTimeout(() => {
-        removeTransition(target);
-
-        // callback
-        if(typeof fn === 'function') fn();
-    }, duration);
-}
-
-export function destroySlide(context){
-    context.wrapper.querySelectorAll(context.options.receiver).forEach(target => {
-        setCSS(target, {
-            display: '',
-            height: ''
-        });
-
-        removeTransition(target);
-    });
-}
 
 /**
  * Fade
@@ -157,7 +82,7 @@ export function fadeIn(target, duration = 500, fn){
  * Helpers
  */
 
-function getElementMarginHeight(el){
+export function getElementMarginHeight(el){
     const computedStyle = getComputedStyle(el);
     return parseInt(computedStyle.getPropertyValue('margin-top') + computedStyle.getPropertyValue('margin-bottom'), 10);
 }
@@ -166,7 +91,7 @@ export function setCSS(target, props){
     Object.assign(target.style, props);
 }
 
-function setTransition(target, duration){
+export function setTransition(target, duration){
     setCSS(target, {
         transitionProperty: "height, margin, padding, opacity",
         transitionDuration: duration + 'ms',
@@ -174,7 +99,7 @@ function setTransition(target, duration){
     });
 }
 
-function removeTransition(target){
+export function removeTransition(target){
     setCSS(target, {
         overflow: '',
         transitionProperty: '',
